backend/requirements.txt
# ShiftWise - Requirements
# Backend dependencies for intelligent VM migration platform

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# API Documentation
uvicorn==0.27.0

# Utilities
python-dateutil==2.8.2

# Hypervisor APIs (for future implementation)
pyvmomi==8.0.1.0.2
libvirt-python==10.0.0; platform_system == "Linux"

# Kubernetes & OpenShift (for future implementation)
kubernetes==29.0.0

# File operations & conversions (for future implementation)
paramiko==3.4.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0